코드 리팩토링 및 테스트 추가
- Controller
- Service
- Entity 및 Service 테스트

+++ src/main/java/com/se/fileserver/v1/file/application/service/FileDeleteService.java
@@ -1,7 +1,7 @@
 package com.se.fileserver.v1.file.application.service;
 
-import com.se.fileserver.v1.common.domain.exception.BusinessException;
-import com.se.fileserver.v1.file.application.service.error.FileErrorCode;
+import com.se.fileserver.v1.common.domain.exception.NotFoundException;
+import com.se.fileserver.v1.file.domain.exception.FileDeletionException;
 import com.se.fileserver.v1.file.domain.model.File;
 import com.se.fileserver.v1.file.domain.repository.FileRepositoryProtocol;
 import java.nio.file.Files;
@@ -12,45 +12,43 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class FileDeleteService {
+
   private final Path fileLocation;
 
   private final FileRepositoryProtocol fileRepositoryProtocol;
 
   public FileDeleteService(@Value("${se-file-server.upload-dir}") String directory,
       FileRepositoryProtocol fileRepositoryProtocol) {
-    this.fileLocation = Paths.get(directory).toAbsolutePath().normalize();
     this.fileRepositoryProtocol = fileRepositoryProtocol;
-
-    try {
-      Files.exists(this.fileLocation);
-    }
-    catch(Exception e) {
-      // TODO: 파일 삭제에 대한 에러코드 정의
-      throw new BusinessException(FileErrorCode.DOWNLOAD_PATH_DOES_NOT_EXISTS);
-    }
+    this.fileLocation = ensureDownloadDir(directory);
   }
 
-  public String deleteFile(String saveName, String service){
+  public void deleteFile(String service, String saveName) {
     Path filePath = this.fileLocation.resolve(service).resolve(saveName).normalize();
     java.io.File file = new java.io.File(filePath.toString());
+    File fileModel = fileRepositoryProtocol.findBySaveName(saveName);
 
-    if (!file.exists()) {
-      throw new BusinessException(FileErrorCode.FILE_DOES_NOT_EXISTS);
+    if (!file.exists() || fileModel == null) {
+      throw new NotFoundException("존재하지 않는 파일입니다.");
     }
 
     if (!file.delete()) {
-      // 파일 삭제 실패했다는 enum 필요
-      throw new BusinessException(FileErrorCode.FILE_DOES_NOT_EXISTS);
-    }
-
-    File fileModel = fileRepositoryProtocol.findBySaveName(saveName);
-
-    if (fileModel == null) {
-      throw new BusinessException(FileErrorCode.FILE_DOES_NOT_EXISTS);
+      throw new FileDeletionException("파일 삭제 중 오류가 발생했습니다.");
     }
 
     fileRepositoryProtocol.delete(fileModel);
-    // TODO: 파일 삭제 리턴 타입 정하기
-    return "Deleted successfully.";
+  }
+
+  private Path ensureDownloadDir(String directory) {
+    Path fileLocation;
+    try {
+      fileLocation = Paths.get(directory).toAbsolutePath().normalize();
+      if (!Files.exists(fileLocation)) {
+        throw new Exception();
+      }
+    } catch (Exception e) {
+      throw new NotFoundException("존재하지 않는 파일 경로입니다.");
+    }
+    return fileLocation;
   }
 }
